refactor(security): add auth controler in security modules
