+++ core-server/app/src/main/java/fr/revoicechat/core/nls/UserErrorCode.java
@@ -6,7 +6,6 @@ public enum UserErrorCode implements ErrorLocalizedMessage {
   USER_NOT_FOUND,
   USER_LOGIN_INVALID,
   USER_WITH_NO_VALID_INVITATION,
-  USER_PASSWORD_INVALID,
   USER_PASSWORD_WRONG,
   USER_PASSWORD_WRONG_CONFIRMATION,
 }
