+++ core-server/app/src/main/java/fr/revoicechat/core/repository/UserRepository.java
@@ -6,7 +6,6 @@ import java.util.stream.Stream;
 import fr.revoicechat.core.model.User;
 
 public interface UserRepository {
-  User findByLogin(String login);
   Stream<User> findByServers(UUID serverID);
   long count();
   Stream<User> everyone();
