+++ core-server/app/src/test/java/fr/revoicechat/core/web/TestAuthController.java
@@ -9,9 +9,8 @@ import org.junit.jupiter.api.Test;
 import fr.revoicechat.core.junit.CleanDatabase;
 import fr.revoicechat.core.quarkus.profile.BasicIntegrationTestProfile;
 import fr.revoicechat.core.representation.NewUserRepresentation;
-import fr.revoicechat.core.technicaldata.login.UserPassword;
+import fr.revoicechat.security.representation.UserPassword;
 import fr.revoicechat.core.technicaldata.user.NewUserSignup;
-import fr.revoicechat.core.representation.UserRepresentation;
 import fr.revoicechat.notification.model.ActiveStatus;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
