+++ core-server/live/src/test/resources/application.properties
@@ -4,3 +4,9 @@
 %test.quarkus.datasource.jdbc.url=jdbc:h2:mem:test;DB_CLOSE_DELAY=-1
 %test.quarkus.datasource.username=sa
 %test.quarkus.datasource.password=sa
+
+revoicechat.global.password.min-length=0
+revoicechat.global.password.min-uppercase=0
+revoicechat.global.password.min-lowercase=0
+revoicechat.global.password.min-number=0
+revoicechat.global.password.min-special-char=0
