+++ core-server/security/src/main/java/fr/revoicechat/security/nls/UserErrorCode.java
@@ -4,4 +4,6 @@ import fr.revoicechat.i18n.ErrorLocalizedMessage;
 
 public enum UserErrorCode implements ErrorLocalizedMessage {
   USER_NOT_FOUND,
+  USER_PASSWORD_INVALID,
+  USER_PASSWORD_WRONG_CONFIRMATION,
 }
