+++ core-server/security/src/test/java/fr/revoicechat/security/service/TestJwtService.java
@@ -50,5 +50,7 @@ class TestJwtService {
       String getDisplayName,
       String getLogin,
       Set<String> getRoles
-  ) implements AuthenticatedUser {}
+  ) implements AuthenticatedUser {
+    @Override public String getPassword() {return "";}
+  }
 }
